review: internalize helper, close mixed owned→bare ceiling gap, fix 'bare' terminology
- Move countStepStartedEvents to an internal module (src/runtime/
  count-step-started-events.ts) instead of exporting it from the public
  ./runtime subpath; tests import it relatively.
- Replace the background ceiling's 'unowned' scope with 'totalAttempts'
  (bare starts + largest single owner's starts), so a step that burns part
  of its retry budget under inline owned recovery and then transitions to
  queued/bare retries still trips the combined maxRetries ceiling, while
  racers' one-off stamped duplicates still don't accumulate. Adds the
  mixed-sequence regression test from review.
- Fix comments calling the owned-recovery start a 'bare step_started' —
  it carries the ownerMessageId stamp; only the background path's start
  is bare.
